refactor: resolve final clippy warnings - param reduction, allow attrs, cast cleanup
- Remove panel_bg/cell_border/border_color params from render_session_cell_with_terminal,
  derive them from theme internally; update both call sites and remove now-unused
  variables (primary_color, cell_border, cell_height) from callers
- Remove cell_height param from render_split_node (leaves never used it, recurse passed None)
- Add #[allow(clippy::too_many_arguments)] to aligned_icon_label_row,
  aligned_icon_label_row_with_offset, and build_priority_button (many call sites)
- Remove redundant `as u8` cast on `code + 32` (code is already u8)

cargo clippy --all-features -- -D warnings: clean
cargo test --all: all pass

+++ crates/codirigent-ui/src/input/mouse.rs
@@ -142,7 +142,7 @@ impl TerminalMouseEvent {
                     0x1b,
                     b'[',
                     b'M',
-                    (code + 32) as u8,
+                    code + 32,
                     (self.col + 1 + 32) as u8,
                     (self.row + 1 + 32) as u8,
                 ])
